# Add JRuby support + ASCII optimization