--- a/fast_blank.gemspec
+++ b/fast_blank.gemspec
@@ -10,19 +10,25 @@ Gem::Specification.new do |s|
   s.homepage = 'https://github.com/SamSaffron/fast_blank'
   s.license = 'MIT'
 
-  s.extensions = ['ext/fast_blank/extconf.rb']
   s.require_paths = ['lib']
   s.files = [
     'MIT-LICENSE',
     'README.md',
     'benchmark',
     'lib/.gemkeep',
-    'ext/fast_blank/fast_blank.c',
-    'ext/fast_blank/extconf.rb',
+    'lib/fast_blank.rb',
   ]
   s.test_files = Dir['spec/**/*_spec.rb']
 
-  s.platform = Gem::Platform::RUBY
+  if RUBY_ENGINE == 'jruby'
+    s.platform = 'java'
+    s.files += Dir['lib/**/*.jar'] + Dir['ext/**/*.java']
+  else
+    s.platform = Gem::Platform::RUBY
+    s.extensions = ['ext/fast_blank/extconf.rb']
+    s.files += %w[ext/fast_blank/fast_blank.c ext/fast_blank/extconf.rb]
+  end
+
   s.rubygems_version = '1.3.7'
 
   s.add_development_dependency 'rake-compiler'