… lib/fast_blank.jar