I actually did two things:
  1. Update build files
    a. to be able to 'rake compile' to regenerate the .jar file (which you should not need to do hopefully)
    b. Allow it so 'jruby -S rake build' will make fast_blank-java version of gem.
  2. Fast path common 7-bit ASCII vs always codepoint walk the string

I was not sure if or how I should document the release process but it is to use MRI and JRuby and in both generate a gem file.  When done with JRuby it will generate the -java version of the gem.  Compiling should never be needed unless we discover a bug.  I still added the ability to recompile anyways if the need does arise.

Here is the results of implementing the 7-bit fast path.  The C extension should see a similar benefit doing the same thing (MRI also have a CR_7BIT flag).

BEFORE:
```text
Ignoring bcrypt-ruby-3.0.0-java because its extensions are not built. Try: gem pristine bcrypt-ruby --version 3.0.0

================== Test String Length: 0 ==================
Warming up --------------------------------------
          Fast Blank     5.332M i/100ms
  Fast ActiveSupport     5.301M i/100ms
          Slow Blank   563.513k i/100ms
      New Slow Blank     3.352M i/100ms
Calculating -------------------------------------
          Fast Blank     55.271M (± 8.2%) i/s -    277.239M in   5.052431s
  Fast ActiveSupport     43.290M (± 3.1%) i/s -    217.355M in   5.026307s
          Slow Blank      4.551M (± 3.3%) i/s -     23.104M in   5.082553s
      New Slow Blank     29.058M (± 2.1%) i/s -    147.494M in   5.078192s

Comparison:
          Fast Blank: 55271026.9 i/s
  Fast ActiveSupport: 43289629.3 i/s - 1.28x  (± 0.00) slower
      New Slow Blank: 29057988.5 i/s - 1.90x  (± 0.00) slower
          Slow Blank:  4550890.6 i/s - 12.15x  (± 0.00) slower

================== Test String Length: 6 ==================
Warming up --------------------------------------
          Fast Blank     1.117M i/100ms
  Fast ActiveSupport     1.043M i/100ms
          Slow Blank   246.967k i/100ms
      New Slow Blank   696.700k i/100ms
Calculating -------------------------------------
          Fast Blank     11.482M (± 2.2%) i/s -     58.087M in   5.061280s
  Fast ActiveSupport     10.271M (± 2.0%) i/s -     52.160M in   5.080475s
          Slow Blank      2.458M (± 2.7%) i/s -     12.348M in   5.027789s
      New Slow Blank      7.123M (± 3.0%) i/s -     36.228M in   5.090551s

Comparison:
          Fast Blank: 11482301.1 i/s
  Fast ActiveSupport: 10271006.8 i/s - 1.12x  (± 0.00) slower
      New Slow Blank:  7123466.5 i/s - 1.61x  (± 0.00) slower
          Slow Blank:  2457953.2 i/s - 4.67x  (± 0.00) slower

================== Test String Length: 14 ==================
Warming up --------------------------------------
          Fast Blank     2.626M i/100ms
  Fast ActiveSupport     2.746M i/100ms
          Slow Blank   468.750k i/100ms
      New Slow Blank   491.832k i/100ms
Calculating -------------------------------------
          Fast Blank     28.403M (± 3.4%) i/s -    144.414M in   5.090509s
  Fast ActiveSupport     27.535M (± 2.4%) i/s -    140.059M in   5.089567s
          Slow Blank      4.730M (± 2.0%) i/s -     23.906M in   5.056373s
      New Slow Blank      4.975M (± 2.4%) i/s -     25.083M in   5.045453s

Comparison:
          Fast Blank: 28403306.0 i/s
  Fast ActiveSupport: 27535241.3 i/s - same-ish: difference falls within error
      New Slow Blank:  4974523.1 i/s - 5.71x  (± 0.00) slower
          Slow Blank:  4729839.9 i/s - 6.01x  (± 0.00) slower

================== Test String Length: 24 ==================
Warming up --------------------------------------
          Fast Blank     1.328M i/100ms
  Fast ActiveSupport     1.385M i/100ms
          Slow Blank   291.609k i/100ms
      New Slow Blank   504.248k i/100ms
Calculating -------------------------------------
          Fast Blank     13.549M (± 2.1%) i/s -     67.740M in   5.001784s
  Fast ActiveSupport     13.937M (± 2.4%) i/s -     70.628M in   5.070878s
          Slow Blank      2.788M (± 3.5%) i/s -     13.997M in   5.027617s
      New Slow Blank      4.906M (± 2.6%) i/s -     24.708M in   5.039930s

Comparison:
  Fast ActiveSupport: 13936752.4 i/s
          Fast Blank: 13549334.5 i/s - same-ish: difference falls within error
      New Slow Blank:  4905915.6 i/s - 2.84x  (± 0.00) slower
          Slow Blank:  2787666.7 i/s - 5.00x  (± 0.00) slower

================== Test String Length: 136 ==================
Warming up --------------------------------------
          Fast Blank     1.326M i/100ms
  Fast ActiveSupport     1.377M i/100ms
          Slow Blank   274.371k i/100ms
      New Slow Blank   487.043k i/100ms
Calculating -------------------------------------
          Fast Blank     13.213M (± 4.5%) i/s -     66.319M in   5.030330s
  Fast ActiveSupport     13.622M (± 4.4%) i/s -     68.854M in   5.065124s
          Slow Blank      2.818M (± 3.5%) i/s -     14.267M in   5.068734s
      New Slow Blank      4.961M (± 2.6%) i/s -     24.839M in   5.009949s

Comparison:
  Fast ActiveSupport: 13621613.2 i/s
          Fast Blank: 13212879.5 i/s - same-ish: difference falls within error
      New Slow Blank:  4961429.0 i/s - 2.75x  (± 0.00) slower
          Slow Blank:  2818282.8 i/s - 4.83x  (± 0.00) slower
```

AFTER:
```text
Ignoring bcrypt-ruby-3.0.0-java because its extensions are not built. Try: gem pristine bcrypt-ruby --version 3.0.0

================== Test String Length: 0 ==================
Warming up --------------------------------------
          Fast Blank     5.183M i/100ms
  Fast ActiveSupport     5.475M i/100ms
          Slow Blank   555.604k i/100ms
      New Slow Blank     3.747M i/100ms
Calculating -------------------------------------
          Fast Blank     60.114M (± 4.8%) i/s -    300.627M in   5.013318s
  Fast ActiveSupport     55.138M (± 9.3%) i/s -    279.203M in   5.110078s
          Slow Blank      4.490M (± 3.3%) i/s -     22.780M in   5.079362s
      New Slow Blank     29.984M (± 2.1%) i/s -    149.872M in   5.000643s

Comparison:
          Fast Blank: 60113553.9 i/s
  Fast ActiveSupport: 55137618.9 i/s - same-ish: difference falls within error
      New Slow Blank: 29984326.8 i/s - 2.00x  (± 0.00) slower
          Slow Blank:  4489941.0 i/s - 13.39x  (± 0.00) slower

================== Test String Length: 6 ==================
Warming up --------------------------------------
          Fast Blank     3.277M i/100ms
  Fast ActiveSupport     1.044M i/100ms
          Slow Blank   238.001k i/100ms
      New Slow Blank   740.737k i/100ms
Calculating -------------------------------------
          Fast Blank     35.563M (± 3.1%) i/s -    180.208M in   5.072370s
  Fast ActiveSupport     10.757M (± 2.0%) i/s -     54.310M in   5.051226s
          Slow Blank      2.303M (± 4.5%) i/s -     11.662M in   5.073777s
      New Slow Blank      7.376M (± 5.5%) i/s -     37.037M in   5.037179s

Comparison:
          Fast Blank: 35563124.2 i/s
  Fast ActiveSupport: 10756502.4 i/s - 3.31x  (± 0.00) slower
      New Slow Blank:  7375883.1 i/s - 4.82x  (± 0.00) slower
          Slow Blank:  2303423.4 i/s - 15.44x  (± 0.00) slower

================== Test String Length: 14 ==================
Warming up --------------------------------------
          Fast Blank     4.405M i/100ms
  Fast ActiveSupport     2.855M i/100ms
          Slow Blank   469.881k i/100ms
      New Slow Blank   459.742k i/100ms
Calculating -------------------------------------
          Fast Blank     42.903M (± 5.2%) i/s -    215.846M in   5.045625s
  Fast ActiveSupport     28.698M (± 1.9%) i/s -    145.591M in   5.075173s
          Slow Blank      4.673M (± 2.8%) i/s -     23.494M in   5.032132s
      New Slow Blank      4.883M (± 3.2%) i/s -     24.826M in   5.090074s

Comparison:
          Fast Blank: 42903124.8 i/s
  Fast ActiveSupport: 28697707.5 i/s - 1.50x  (± 0.00) slower
      New Slow Blank:  4882707.8 i/s - 8.79x  (± 0.00) slower
          Slow Blank:  4672514.2 i/s - 9.18x  (± 0.00) slower

================== Test String Length: 24 ==================
Warming up --------------------------------------
          Fast Blank     4.103M i/100ms
  Fast ActiveSupport     1.432M i/100ms
          Slow Blank   288.582k i/100ms
      New Slow Blank   473.903k i/100ms
Calculating -------------------------------------
          Fast Blank     41.346M (± 2.0%) i/s -    209.278M in   5.063666s
  Fast ActiveSupport     14.514M (± 2.0%) i/s -     73.021M in   5.033085s
          Slow Blank      2.771M (± 5.2%) i/s -     13.852M in   5.013736s
      New Slow Blank      4.816M (± 2.6%) i/s -     24.169M in   5.021804s

Comparison:
          Fast Blank: 41345655.1 i/s
  Fast ActiveSupport: 14514324.0 i/s - 2.85x  (± 0.00) slower
      New Slow Blank:  4816073.0 i/s - 8.58x  (± 0.00) slower
          Slow Blank:  2771099.9 i/s - 14.92x  (± 0.00) slower

================== Test String Length: 136 ==================
Warming up --------------------------------------
          Fast Blank     3.751M i/100ms
  Fast ActiveSupport     1.382M i/100ms
          Slow Blank   289.175k i/100ms
      New Slow Blank   473.207k i/100ms
Calculating -------------------------------------
          Fast Blank     40.670M (± 3.8%) i/s -    206.325M in   5.081474s
  Fast ActiveSupport     14.307M (± 2.5%) i/s -     71.874M in   5.026807s
          Slow Blank      2.857M (± 2.5%) i/s -     14.459M in   5.064108s
      New Slow Blank      4.828M (± 2.8%) i/s -     24.134M in   5.002497s

Comparison:
          Fast Blank: 40669856.0 i/s
  Fast ActiveSupport: 14307278.3 i/s - 2.84x  (± 0.00) slower
      New Slow Blank:  4828182.4 i/s - 8.42x  (± 0.00) slower
          Slow Blank:  2857047.0 i/s - 14.23x  (± 0.00) slower
```