@SamSaffron 2021-12-03
Awesome @enebo! 

Will have a look at publishing a new version of the gem next week. Some Java instructions would certainly be helpful.