@SamSaffron 2021-12-06
I see, was pretty simple, I just pushed 1.0.1 (Java edition) is it working
well for you?

On Sat, Dec 4, 2021 at 1:14 AM Thomas E Enebo ***@***.***>
wrote:

> @SamSaffron <https://github.com/SamSaffron> I can make a PR in the right
> place or I can make a little section here and you can commit it somewhere.
> If you run into problems or have any questions I will try and pay attention
> to any pings but sometimes I miss stuff on github. If I don't respond
> quickly you can message me on twittter ***@***.***_enebo).
>
> If this had been a more complicated gem I would have changed this to
> require compiling fast_blank every time you release but this puts a burden
> on needing a java compiler to perform a release. This gem is pretty minimal.
> JRuby Support
>
> To publish the JRuby version of this gem you need to just run the same
> command you would run to publish the CRuby version but you need to run it
> with jruby (e.g. 'jruby -S rake build'). This will make a second version of
> the gem which a JRuby user will get when they install this gem.
>
> The main differences between the two versions of the gem is the JRuby
> version will not do a compile as part of the creation (since a compiled
> version of the native bits has already been committed). The JRuby gem will
> have a platform of 'java' and will include the Java bits instead of the C
> bits (also excluding ext_conf.rb -- since it is already compiled).
>
> In the rare case that we need to recompile the Java bits you can run an
> explcit 'rake compile' using JRuby and it will create a new version of
> lib/fast_blank.jar. After running this compile commit the new .jar file and
> then do a new release. To perform this you will need a version of JDK (Java
> Development Kit) version 1.8. You can also open an issue on the JRuby
> project asking for a new compile. We will open a PR with the new bits.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/SamSaffron/fast_blank/pull/31#issuecomment-985553904>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAABIXLAGUQK27763Z6CGL3UPDGDBANCNFSM5JIBXDFQ>
> .
>